# swagger api-plantilla backend.yaml
openapi: 3.0.3
info:
  title: Clerkiva API
  description: API for Clerkiva backend services.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "Recurso no encontrado"
        error:
          type: string
          example: "Not Found"
    Servicio:
      type: object
      properties:
        idServicio:
          type: integer
          example: 1
        servicio:
          type: string
          maxLength: 60
          example: "Corte de pelo"
        duracion:
          type: integer
          description: Duración del turno en minutos
          example: 30
        precio:
          type: number
          example: 8000
        estado:
          type: string
          description: Estado del servicio (A=Activo, B=Baja)
          enum: [A, B]
          example: "A"
    ServicioInput:
      type: object
      required:
        - servicio
        - duracion
        - precio
      properties:
        servicio:
          type: string
          maxLength: 60
          example: "Corte de pelo"
        duracion:
          type: integer
          description: Duración del turno en minutos
          example: 30
        precio:
          type: number
          example: 8000
        estado:
          type: string
          description: Si no se envía, se guarda como A
          enum: [A, B]
          example: "A"
...
paths:
  /auth:
    get:
      tags:
        - Auth
      summary: Obtener usuario autenticado
      description: Devuelve la información del usuario dueño del JWT. Requiere token válido en el header Authorization.
      operationId: getAuth
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: Token de acceso obtenido al iniciar sesión. Debe ser enviado en el formato "Bearer {token}"
      responses:
        "200":
          description: Datos del usuario autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  idUsuario:
                    type: integer
                    description: ID del usuario autenticado
                    example: 1
                  usuario:
                    type: string
                    description: Nombre de usuario del usuario autenticado
                    example: "feder"
        "401":
          description: Token inválido o no enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: "Unauthorized"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión. Ingresar Usuario y contraseña y devuelve Token
      description: Autentica al usuario con su usuario y contraseña. Devuelve un JWT que debe usarse en los endpoints protegidos.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                  description: Nombre de usuario del usuario que intenta iniciar sesión
                  example: "admin"
                password:
                  type: string
                  description: Contraseña del usuario que intenta iniciar sesión
                  example: "admin12345"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token de acceso
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkFkbWluIiwiaWF0IjoxNT"
        "400":
          description: Datos de entrada inválidos (validación fallida)
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Invalid input"
        "401":
          description: Contraseña incorrecta, o usuario dado de baja (estado distinto de 'A')
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: "Invalid credentials"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "User not found"
...
  /servicios:
    get:
      tags:
        - Servicios
      summary: Obtener todos los servicios
      description: Devuelve todos los servicios (activos y dados de baja) ordenados por nombre
      operationId: getAllServicios
      responses:
        "200":
          description: Listado de servicios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Servicio"
    post:
      tags:
        - Servicios
      summary: Alta de servicio
      operationId: createServicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServicioInput"
      responses:
        "201":
          description: Servicio creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servicio"
  /servicios/nombre/{nombre}:
    get:
      tags:
        - Servicios
      summary: Buscar servicios activos por nombre
      description: Devuelve los servicios activos cuyo nombre contiene el texto (sin distinguir mayúsculas), del más barato al más caro. No incluye el campo estado.
      operationId: getServiciosByNombre
      parameters:
        - in: path
          name: nombre
          required: true
          schema:
            type: string
          example: "corte"
      responses:
        "200":
          description: Servicios encontrados (puede ser un array vacío)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Servicio"
  /servicios/{id}:
    get:
      tags:
        - Servicios
      summary: Obtener un servicio
      operationId: getServicioById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del servicio
          example: 1
      responses:
        "200":
          description: Servicio encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servicio"
        "404":
          description: Servicio no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 404
                message: "El Servicio 1 no fue encontrado"
                error: "Not Found"
    put:
      tags:
        - Servicios
      summary: Reemplazar un servicio
      description: Reemplaza todos los campos del servicio
      operationId: replaceServicio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del servicio
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServicioInput"
      responses:
        "200":
          description: Servicio actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servicio"
        "404":
          description: Servicio no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 404
                message: "El Servicio 1 no fue encontrado"
                error: "Not Found"
    patch:
      tags:
        - Servicios
      summary: Modificar parcialmente un servicio
      description: Modifica solo los campos enviados
      operationId: updateServicio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del servicio
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                servicio:
                  type: string
                duracion:
                  type: integer
                precio:
                  type: number
                estado:
                  type: string
                  enum: [A, B]
            example:
              precio: 9500
      responses:
        "200":
          description: Servicio actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servicio"
        "404":
          description: Servicio no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 404
                message: "El Servicio 1 no fue encontrado"
                error: "Not Found"
    delete:
      tags:
        - Servicios
      summary: Dar de baja un servicio
      description: Baja lógica, pasa el estado a B. El servicio no se borra porque puede tener turnos asociados.
      operationId: deleteServicio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del servicio
          example: 1
      responses:
        "200":
          description: Servicio dado de baja
        "404":
          description: Servicio no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 404
                message: "El Servicio 1 no fue encontrado"
                error: "Not Found"
...